+++ SSD1306_AVR/twi.h
@@ -6,7 +6,7 @@
 void TWI_Init(void);
 void TWI_Start(void);
 void TWI_Stop(void);
-void TWI_SendByte(int data);
-void TWI_SendAddress(int addr);
+void TWI_SendByte(uint8_t data);
+void TWI_SendAddress(uint8_t addr);
 
 #endif
